+++ packages/expo-haptics/src/Haptics.ts
@@ -3,28 +3,6 @@ import { UnavailabilityError } from '@unimodules/core';
 import ExpoHaptics from './ExpoHaptics';
 import { NotificationFeedbackType, ImpactFeedbackStyle } from './Haptics.types';
 
-/** @deprecated Use `Haptics.notificationAsync` instead. */
-export async function notification(
-  type: NotificationFeedbackType = NotificationFeedbackType.Success
-): Promise<void> {
-  console.warn('`Haptics.notification` is deprecated. Use `Haptics.notificationAsync` instead.');
-  await notificationAsync(type);
-}
-
-/** @deprecated Use `Haptics.impactAsync` instead */
-export async function impact(
-  style: ImpactFeedbackStyle = ImpactFeedbackStyle.Medium
-): Promise<void> {
-  console.warn('`Haptics.impact` is deprecated. Use `Haptics.impactAsync` instead.');
-  await impactAsync(style);
-}
-
-/** @deprecated Use `Haptics.selectionAsync` instead */
-export async function selection(): Promise<void> {
-  console.warn('`Haptics.selection` is deprecated. Use `Haptics.selectionAsync` instead.');
-  await selectionAsync();
-}
-
 /**
  * Triggers notification feedback.
  */